+++ auth-service/src/main/java/com/ecommerce/auth/service/AdminService.java
@@ -1,12 +1,14 @@
 package com.ecommerce.auth.service;
 
 
+import com.ecommerce.auth.dto.PaginatedResponse;
 import com.ecommerce.auth.dto.UserDto;
 
 import java.util.List;
 
 public interface AdminService {
-    List<UserDto> getAllUsers();
+    PaginatedResponse<UserDto> getAllUsers(int page, int size);
+
     UserDto getUserById(Long id);
     void deleteUserById(Long id);
 }